move files in proper folders and remove redundant code
