+++ frontend/src/components/servers/PlayersCard.jsx
@@ -1,7 +1,7 @@
 import { X } from 'lucide-react';
 import PropTypes from 'prop-types';
 import PlayerList from './PlayerList';
-import { Button } from './ui/button';
+import { Button } from '../ui/button';
 
 export const PlayersCard = ({ players, togglePlayerList }) => (
   <article className="w-[318px] flex flex-col text-white h-full p-4 bg-slate-700 shadow-lg rounded overflow-hidden border-2 border-gray-500 hover:border-gray-300 cursor-pointer">
